+++ vueui/src/components/view/records/field.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-import Utils from '../../../utils.js'
+import Utils from '../../../vueuiutils.js'
 export default {
   data() {
     return {
