+++ vueui/src/components/view/records/medical.vue
@@ -23,7 +23,7 @@
 </template>
 
 <script>
-import Utils from '../../../utils.js'
+import Utils from '../../../vueuiutils.js'
 export default {
   data() {
     return this.$root.$data.state
