+++ vueui/src/components/view/records/security.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script>
-import Utils from '../../../utils.js'
+import Utils from '../../../vueuiutils.js'
 export default {
   data() {
     return this.$root.$data.state
